+++ Changelog.md
@@ -1,6 +1,7 @@
 # Next
 
 - **Breaking Change** Throw dedicated `Error.jsonMapping` when `mapJSON` fails to parse JSON
+- Removed the unused `StreamRequest` typealias that was causing watchOS failures.
 
 # 8.0.0-beta.1